Minor bugs fixes and restructure of the directories and files, also now it compiles with the use of a Makefile
